use tgstat functions bug fix: import Matrix classes and as.tibble

+++ R/smat_import.R
@@ -11,18 +11,26 @@
 #' @param use_sge use sun grid engine cluster
 #' @param name name of the object
 #' @param description description of the object
+#' @param keep_tidy_cpgs keep tidy cpgs attached to the smat object
 #' @param ... additional parameters to \code{gpatterns::gpatterns.import_from_bam}
 #' 
 #' @return smat object (inivisibly if prefix is not NULL)
 #'
 #' @export
-smat.from_bams <- function(metadata, groot, prefix=NULL, workdir=tempdir(), use_sge = TRUE, name='', description='', ...){
+smat.from_bams <- function(metadata, groot, prefix=NULL, workdir=tempdir(), use_sge = TRUE, name='', description='', keep_tidy_cpgs=FALSE, load_existing=FALSE, io_saturation=TRUE, ...){
 
-    bam2calls <- function(bams, lib, workdir=workdir, use_sge=TRUE){                  
-        track_workdir <- tempfile(tmpdir=workdir)
+    bam2calls <- function(bams, lib, workdir=workdir, ...){                  
+        if (keep_tidy_cpgs){
+            track_workdir <- paste0(prefix, '_tcpgs/', lib)            
+        } else {
+            track_workdir <- tempfile(tmpdir=workdir)            
+            on.exit(system(sprintf('rm -rf %s', track_workdir)))            
+        }
         system(sprintf('mkdir -p %s', track_workdir))
-        on.exit(system(sprintf('rm -rf %s', track_workdir)))        
-        gpatterns::gpatterns.import_from_bam(bams, workdir=track_workdir, steps=c('bam2tidy_cpgs', 'filter_dups'), groot=groot)
+        
+        if (!load_existing){
+            gpatterns::gpatterns.import_from_bam(bams, workdir=track_workdir, steps=c('bam2tidy_cpgs', 'filter_dups'), groot=groot, ...)    
+        }        
 
         tcpgs_dir <- paste0(track_workdir, '/tidy_cpgs')
         uniq_tcpgs_dir <- paste0(track_workdir, '/tidy_cpgs_uniq')
@@ -45,24 +53,28 @@ smat.from_bams <- function(metadata, groot, prefix=NULL, workdir=tempdir(), use_
 
     cmds <- plyr::daply(metadata, .(lib), function(x) {
         bams <- paste(qqv("'@{x$bam}'"), collapse=', ')
-        qq("bam2calls(c(@{bams}), lib = '@{x$lib[1]}', workdir='@{workdir}')")
+        qq("bam2calls(c(@{bams}), lib = '@{x$lib[1]}', workdir='@{workdir}', ...)")
     })
     
     if (use_sge){        
-        res <- gcluster.run2(command_list=cmds, io_saturation=T, packages='gpatterns')
+        res <- gcluster.run2(command_list=cmds, io_saturation=io_saturation, packages='gpatterns')
         tidy_calls <- res %>% map('retv') %>% map('calls') %>% compact() %>% map_df(~ .x)
         stats <- res %>% map('retv') %>% map('stats') %>% compact() %>% map_df(~ .x)
     } else {
         tidy_calls <- map(cmds, ~ eval(parse(text=.x))) %>% map('calls') %>% compact() %>% map_df(~ .x)
         stats <- map(cmds, ~ eval(parse(text=.x))) %>% map('stats') %>% compact() %>% map_df(~ .x)
     }
-
+    browser()
     tidy_calls <- tidy_calls %>% mutate(unmeth = if_else(meth == 0, 1, 0), cov=1)
-
-    smat <- .tidy_calls2smat(tidy_calls %>% ungroup() %>% mutate(id = 1:n()))
+    
+    smat <- .tidy_calls2smat(tidy_calls)
     smat$stats <- stats
     smat$name <- name
     smat$description <- description
+
+    if (keep_tidy_cpgs){
+        smat$tidy_cpgs <- paste0(prefix, '_tcpgs') 
+    }
 
     if (!is.null(prefix)){
         system(qq('mkdir -p @{dirname(prefix)}'))
@@ -138,27 +150,22 @@ smat.from_df <- function(df, name='', description=''){
 # Utils
 
 .tidy_calls2smat <- function(tidy_calls, column_name='track'){     
-    tidy_calls <- tidy_calls %>% mutate(id = as.character(id))
+    message('creating intervs')  
+    intervs <- tidy_calls %>% distinct(chrom, start, end) %>% mutate(id = 1:n())
+    
+    if (has_name(tidy_calls, 'id')){
+        tidy_calls <- tidy_calls %>% select(-id)
+    }
+    tidy_calls <- tidy_calls %>% left_join(intervs, by=c('chrom', 'start', 'end'))
 
     smat <- plyr::alply(c('meth', 'unmeth', 'cov'), 1, function(x) {
         message(sprintf('creating %s', x))
         tidy2smat(tidy_calls, 'id', column_name, x)
-    }, .parallel=TRUE)
+    }, .parallel=FALSE)
 
     names(smat) <- c('meth', 'unmeth', 'cov')
-
-    message('creating intervs')    
-    # match intervals and ids
-    smat$intervs <- tidy_calls %>%        
-        inner_join(tibble(id = rownames(smat[[1]])), by='id') %>%
-        distinct(chrom, start, end, id)
-
-    # generate new ids
-    smat$intervs <- smat$intervs %>% mutate(id = 1:n())   
-    
-    # smat$intervs <- tibble(coord = rownames(smat[[1]])) %>% 
-    #     separate(coord, c('chrom', 'start', 'end')) %>% mutate(id = 1:n()) %>% 
-    #     mutate(start = as.numeric(start), end = as.numeric(end))
+      
+    smat$intervs <- intervs
 
     for (.x in c('meth', 'unmeth', 'cov')){        
         rownames(smat[[.x]]) <- smat$intervs$id
@@ -177,7 +184,7 @@ tidy2smat <- function(data, row, column, value, ...){
 
     val <- data[[value]]
 
-    Matrix::sparseMatrix(i = i, j = j, x = val, dimnames = list(row_u, col_u), ...)
+    Matrix::sparseMatrix(i = i, j = j, x = val, dimnames = list(as.character(row_u), as.character(col_u)), ...)
 }
 
 tcpgs2calls <- function(tcpgs, track){
